SCI: Fix detection_internal.h table duplication
